requirements.txt
amqp==2.5.2
asgiref==3.2.3
billiard==3.6.1.0
celery==4.3.0
Django==3.0
django-crispy-forms==1.8.1
importlib-metadata==1.1.0
kombu==4.6.6
more-itertools==8.0.0
Pillow==6.2.1
pytz==2019.3
redis==3.3.11
sqlparse==0.3.0
vine==1.3.0
zipp==0.6.0
